requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
jinja2==3.1.2
email-validator==2.0.0.post2
python-dotenv==1.0.0
apscheduler==3.10.4
pytest==7.4.2
httpx==0.24.1
pytest-asyncio==0.21.1
